move hit tests
