+++ BitFaster.Caching.UnitTests/Lfu/ConcurrentLfuTests.cs
@@ -759,126 +759,6 @@ namespace BitFaster.Caching.UnitTests.Lfu
             
         }
 
-        //Elapsed 411.6918ms - 0.0004116918ns/op
-        //Cache hits 1689839 (sampled 16.89839%)
-        //Maintenance ops 31
-        [Fact]
-        public void VerifyHitsWithBackgroundScheduler()
-        {
-            // when running all tests in parallel, sample count drops significantly: set low bar for stability.
-            VerifyHits(iterations: 10_000_000, minSamples: 250_000);
-        }
-
-        //Elapsed 590.8154ms - 0.0005908154ns/op
-        //Cache hits 3441470 (sampled 34.414699999999996%)
-        //Maintenance ops 20
-        [Fact]
-        public void VerifyHitsWithThreadPoolScheduler()
-        {
-            // when running all tests in parallel, sample count drops significantly: set low bar for stability.
-            cache = new ConcurrentLfu<int, int>(1, 20, new ThreadPoolScheduler(), EqualityComparer<int>.Default);
-            VerifyHits(iterations: 10_000_000, minSamples: 500_000);
-        }
-
-        //Elapsed 273.0148ms - 0.0002730148ns/op
-        //Cache hits 0 (sampled 0%)
-        //Maintenance ops 1
-        [Fact]
-        public void VerifyHitsWithNullScheduler()
-        {
-            cache = new ConcurrentLfu<int, int>(1, 20, new NullScheduler(), EqualityComparer<int>.Default);
-            VerifyHits(iterations: 10_000_000, minSamples: -1);
-        }
-
-        //Will drop 78125 reads.
-        //Elapsed 847.5331ms - 0.0008475331ns/op
-        //Cache hits 10000000 (sampled 99.2248062015504%)
-        //Maintenance ops 78126
-        [Fact]
-        public void VerifyHitsWithForegroundScheduler()
-        {
-            cache = new ConcurrentLfu<int, int>(1, 20, new ForegroundScheduler(), EqualityComparer<int>.Default);
-
-            // Note: TryAdd will drop 1 read per full read buffer, since TryAdd will return false
-            // before TryScheduleDrain is called. This serves as sanity check.
-            int iterations = 10_000_000;
-            int dropped = iterations / ConcurrentLfu<int, int>.DefaultBufferSize;
-
-            this.output.WriteLine($"Will drop {dropped} reads.");
-
-            VerifyHits(iterations: iterations + dropped, minSamples: iterations);
-        }
-
-        [Fact]
-        public void VerifyMisses()
-        {
-            cache = new ConcurrentLfu<int, int>(1, 20, new BackgroundThreadScheduler(), EqualityComparer<int>.Default);
-
-            int iterations = 100_000;
-            Func<int, int> func = x => x;
-
-            var start = Stopwatch.GetTimestamp();
-
-            for (int i = 0; i < iterations; i++)
-            {
-                cache.GetOrAdd(i, func);
-            }
-
-            var end = Stopwatch.GetTimestamp();
-
-            cache.DoMaintenance();
-
-            var totalTicks = end - start;
-            var timeMs = ((double)totalTicks / Stopwatch.Frequency) * 1000.0;
-            var timeNs = timeMs / 1_000_000;
-
-            var timePerOp = timeMs / (double)iterations;
-            var samplePercent = this.cache.Metrics.Value.Misses / (double)iterations * 100;
-
-            this.output.WriteLine($"Elapsed {timeMs}ms - {timeNs}ns/op");
-            this.output.WriteLine($"Cache misses {this.cache.Metrics.Value.Misses} (sampled {samplePercent}%)");
-            this.output.WriteLine($"Maintenance ops {this.cache.Scheduler.RunCount}");
-
-            cache.Metrics.Value.Misses.Should().Be(iterations);
-        }
-
-        private void VerifyHits(int iterations, int minSamples)
-        {
-            Func<int, int> func = x => x;
-            cache.GetOrAdd(1, func);
-
-            var start = Stopwatch.GetTimestamp();
-
-            for (int i = 0; i < iterations; i++)
-            {
-                cache.GetOrAdd(1, func);
-            }
-
-            var end = Stopwatch.GetTimestamp();
-
-            var totalTicks = end - start;
-            var timeMs = ((double)totalTicks / Stopwatch.Frequency) * 1000.0;
-            var timeNs = timeMs / 1_000_000;
-
-            var timePerOp = timeMs / (double)iterations;
-            var samplePercent = this.cache.Metrics.Value.Hits / (double)iterations * 100;
-
-            this.output.WriteLine($"Elapsed {timeMs}ms - {timeNs}ns/op");
-            this.output.WriteLine($"Cache hits {this.cache.Metrics.Value.Hits} (sampled {samplePercent}%)");
-            this.output.WriteLine($"Maintenance ops {this.cache.Scheduler.RunCount}");
-
-            if (this.cache.Scheduler.LastException.HasValue)
-            {
-                this.output.WriteLine($"Error: {this.cache.Scheduler.LastException.Value}");
-            }
-
-            cache.Metrics.Value.Hits.Should().BeGreaterThanOrEqualTo(minSamples);
-
-            // verify this doesn't block or throw
-            var b = cache.Scheduler as BackgroundThreadScheduler;
-            b?.Dispose();
-        }
-
         private void LogLru()
         {
 #if DEBUG
